Regen sln whenever a file is added or moved
To make debugging with VisualStudio useful, we need the sln to have all
the latest files.

Don't care so much when files are removed. The sln will be regenerated
on startup and can be regenerated manually by the user.

Keeping the sln up to date ensures Omnisharp is always working with the
right files.